LayoutTests/imported/w3c/web-platform-tests/trusted-types/require-trusted-types-for-report-only-expected.txt

CONSOLE MESSAGE: The Content Security Policy 'require-trusted-types-for 'script'' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.

Harness Error (TIMEOUT), message = null

TIMEOUT Require trusted types for 'script' block create HTML. Test timed out
NOTRUN Require trusted types for 'script' block create script.
NOTRUN Require trusted types for 'script' block create script URL.
NOTRUN Set require trusted types for 'script' without CSP for trusted types don't block policy creation and using.
